feat: implement RegisterUserUseCaseImpl with bcrypt hashing and email uniqueness validation
